+++ src/components/Content/Content.js
@@ -1,25 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { getContent } from '../../utils';
 import './Content.scss';
 
-const Content = ({ title, text, sampleLength }) => (
+const Content = ({ title, text }) => (
   <div className="content">
     {title && <h1 className="uk-heading-divider">{title}</h1>}
-    <div className="text">{text || getContent(sampleLength)}</div>
+    <div className="text">{text}</div>
   </div>
 );
 
 Content.propTypes = {
   title: PropTypes.string,
-  text: PropTypes.string,
-  sampleLength: PropTypes.number
+  text: PropTypes.string.isRequired
 };
 
 Content.defaultProps = {
-  title: '[Content Title]',
-  text: '',
-  sampleLength: 2
+  title: null
 };
 
 export default Content;