+++ src/components/index.js
@@ -7,3 +7,5 @@ export { default as NavBar } from './NavBar';
 export { default as NotificationContainer } from './Notification';
 export { default as Content } from './Content';
 export { default as Counter } from './Counter';
+export { default as List } from './List';
+export { default as ProgressBar } from './ProgressBar';