+++ src/pages/About/About.js
@@ -1,10 +1,13 @@
 import React from 'react';
-import Container from 'components/Container';
-import Content from 'components/Content';
+import {
+  Container,
+  Content
+} from 'components';
+import { getContent } from '../../utils';
 
 const AboutPage = () => (
   <Container size="small">
-    <Content title="About Us" />
+    <Content title="About Us" text={getContent(2)} />
   </Container>
 );