+++ src/pages/Contact/Contact.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import ContactPage from './index';
-
-const wrapper = mount(<ContactPage />);
-
-test('ContactPage Page Renders', () => {
-  expect(wrapper).toMatchSnapshot();
-});