+++ src/pages/Home/Home.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import Container from 'components/Container';
 import Content from 'components/Content';
+import { getContent } from '../../utils';
 
 const HomePage = () => (
   <Container size="small">
-    <Content title="Home" sampleLength={5} />
+    <Content title="Home" text={getContent(2)} />
   </Container>
 );