+++ src/pages/index.js
@@ -2,5 +2,5 @@ export { default as HomePage } from './Home';
 export { default as AboutPage } from './About';
 export { default as ButtonsPage } from './Buttons';
 export { default as CardsPage } from './Cards';
-export { default as ContactPage } from './Contact';
 export { default as TodoPage } from './Todos';
+export { default as ComponentsPage } from './Components';